+++ CHANGES.rst
@@ -2,6 +2,11 @@
 Changes for Lovely Gradle Plugin
 ================================
 
+unreleased
+==========
+
+ - fixed latest local release tag lookup, which failed when only none-release tags are present
+
 2018/03/15 0.0.5
 ================